只有拍照信号和预览信号才会触发下位机通讯

+++ component/include/widgets/doubleslider.h
@@ -26,7 +26,7 @@ public:
     void setSuffix(const QString& s);
     void setSingleStep(int step);
     void setMouseEvent(bool enabled);
-    int value() const;
+    double value() const;
     void setValue(int val);
 
     void addValue(double val);
